Rename common-sbom module to common-maven.
