+++ java-components/build-recipes-database/pom.xml
@@ -19,7 +19,7 @@
         </dependency>
         <dependency>
             <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
-            <artifactId>common-sbom</artifactId>
+            <artifactId>common-maven</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
