+++ java-components/build-recipes-database/src/main/java/com/redhat/hacbs/recipes/scm/AbstractPomScmLocator.java
@@ -12,7 +12,7 @@ import org.apache.maven.model.Parent;
 import org.apache.maven.model.Scm;
 import org.jboss.logging.Logger;
 
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 
 public abstract class AbstractPomScmLocator implements ScmLocator {
