+++ java-components/build-recipes-database/src/main/java/com/redhat/hacbs/recipes/scm/GitScmLocator.java
@@ -21,7 +21,7 @@ import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.lib.Ref;
 import org.jboss.logging.Logger;
 
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 import com.redhat.hacbs.recipes.BuildRecipe;
 import com.redhat.hacbs.recipes.location.RecipeDirectory;
 import com.redhat.hacbs.recipes.location.RecipeGroupManager;
