+++ java-components/build-recipes-database/src/test/java/com/redhat/hacbs/recipes/location/RecipeGroupManagerAddTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.yaml.snakeyaml.Yaml;
 
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 import com.redhat.hacbs.recipes.BuildRecipe;
 import com.redhat.hacbs.recipes.scm.ScmInfo;
