+++ java-components/build-recipes-database/src/test/java/com/redhat/hacbs/recipes/location/RecipeGroupManagerSingleTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 import com.redhat.hacbs.recipes.BuildRecipe;
 import com.redhat.hacbs.recipes.build.BuildRecipeInfo;
 import com.redhat.hacbs.recipes.scm.ScmInfo;
