+++ java-components/build-request-processor/pom.xml
@@ -27,7 +27,7 @@
         </dependency>
         <dependency>
             <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
-            <artifactId>common-sbom</artifactId>
+            <artifactId>common-maven</artifactId>
         </dependency>
 
         <dependency>
