+++ java-components/build-request-processor/src/main/java/com/redhat/hacbs/container/analyser/location/CacheScmLocator.java
@@ -8,7 +8,7 @@ import jakarta.ws.rs.core.MediaType;
 
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 import com.redhat.hacbs.recipes.scm.ScmLocator;
 import com.redhat.hacbs.recipes.scm.TagInfo;
