+++ java-components/build-request-processor/src/main/java/com/redhat/hacbs/container/deploy/containerregistry/ContainerRegistryDeployer.java
@@ -29,7 +29,7 @@ import com.google.cloud.tools.jib.api.buildplan.FileEntriesLayer;
 import com.google.cloud.tools.jib.api.buildplan.FilePermissions;
 import com.google.cloud.tools.jib.api.buildplan.FilePermissionsProvider;
 import com.google.cloud.tools.jib.api.buildplan.ImageFormat;
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 import com.redhat.hacbs.container.deploy.DeployData;
 
 import io.quarkus.logging.Log;
