+++ java-components/cache/src/main/java/com/redhat/hacbs/artifactcache/resources/V2RecipeLookupResource.java
@@ -15,7 +15,7 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 import com.redhat.hacbs.artifactcache.services.RecipeManager;
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 
 import io.quarkus.logging.Log;
 import io.smallrye.common.annotation.Blocking;
