+++ java-components/cache/src/main/java/com/redhat/hacbs/artifactcache/services/RepositoryCache.java
@@ -27,7 +27,7 @@ import org.apache.commons.lang3.StringUtils;
 import com.redhat.hacbs.classfile.tracker.ClassFileTracker;
 import com.redhat.hacbs.classfile.tracker.HashingOutputStream;
 import com.redhat.hacbs.classfile.tracker.TrackingData;
-import com.redhat.hacbs.common.sbom.GAV;
+import com.redhat.hacbs.common.maven.GAV;
 
 import io.quarkus.logging.Log;
