+++ java-components/common-maven/src/main/java/com/redhat/hacbs/common/maven/GAV.java
@@ -1,4 +1,4 @@
-package com.redhat.hacbs.common.sbom;
+package com.redhat.hacbs.common.maven;
 
 import java.util.Comparator;
 import java.util.Objects;
