+++ java-components/domainproxy/server/pom.xml
@@ -30,7 +30,7 @@
         </dependency>
         <dependency>
             <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
-            <artifactId>common-sbom</artifactId>
+            <artifactId>common-maven</artifactId>
         </dependency>
     </dependencies>
