+++ java-components/pom.xml
@@ -82,7 +82,7 @@
         <module>cli</module>
         <module>common-tools</module>
         <module>common-images</module>
-        <module>common-sbom</module>
+        <module>common-maven</module>
         <module>domainproxy</module>
     </modules>
 
@@ -150,7 +150,7 @@
             </dependency>
             <dependency>
                 <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
-                <artifactId>common-sbom</artifactId>
+                <artifactId>common-maven</artifactId>
                 <version>${project.version}</version>
             </dependency>
